Romy - added angle calculation and shi
