+++ src/main/java/edu/greenblitz/robotName/shootingPosition/ShootingPosition.java
@@ -1,6 +1,7 @@
 package edu.greenblitz.robotName.shootingPosition;
 
 import com.sun.source.tree.ReturnTree;
+import edu.greenblitz.robotName.utils.GBCommand;
 import edu.wpi.first.math.geometry.Pose2d;
 
 import java.awt.geom.Ellipse2D;
@@ -8,7 +9,7 @@ import java.security.spec.EllipticCurve;
 
 import static edu.greenblitz.robotName.shootingPosition.ShootingPositionConstants.*;
 
-public class ShootingPosition {
+public class ShootingPosition extends GBCommand {
